Fixed move timeout error on first move

+++ maintainance.py
@@ -1,53 +1,71 @@
+from shared import app
+
 import humanfriendly
 import logging
 from data import User, Match, GameResult
 import constants
+from datetime import datetime, timedelta
 
 
 def task_users():
 
     logging.info("Performing user maintainance...")
 
-    # Delete all users whose last move is older than USER_TIMEOUT
-    query = User.query(User.expired == True)
-    query_list = list(query.fetch())
+    users = list(User.query().fetch())
 
-    for user in query_list:
-        logging.info("Deleting user %s. Last activity: %s", user.username, str(user.last_activity_date))
-        user.key.delete()
+    for user in users:
+        # Delete all users whose last move is older than USER_TIMEOUT
+        expired = (datetime.now() - user.last_activity_date) > timedelta(days=app.config['USER_TIMEOUT'])
+        if expired:
+            # Delete user
+            logging.info("Deleting user %s. Last activity: %s", user.username, str(user.last_activity_date))
+            user.key.delete()
 
 
 def task_matches():
 
     logging.info("Performing match maintainance...")
 
-    # Delete all matches whose last move is older than the specified timeout
-    query = Match.query(Match.expired == True)
-    query_list = list(query.fetch())
+    matches = list(Match.query().fetch())
 
-    for match in query_list:
+    for match in matches:
 
-        logging.info("Deleting match %s. Last activity: %s", match.key.id(), str(match.last_move_date))
-
-        white = User.get_by_id(match.white_id)
-        black = User.get_by_id(match.black_id)
-
-        if match.is_user_turn(white):
-            looser = white
-            winner = black
+        # Delete all matches whose last move is older than the specified timeout
+        if match.last_move_date:
+            expired = (datetime.now() - match.last_move_date) > timedelta(seconds=match.timeout)
         else:
-            winner = white
-            looser = black
+            expired = False
 
-        white.send_message(constants.ERROR_TIMEOUT.format(
-            looser.username,
-            humanfriendly.format_timespan(match.timeout)))
-        black.send_message(constants.ERROR_TIMEOUT.format(
-            looser.username,
-            humanfriendly.format_timespan(match.timeout)))
+        if expired:
 
-        winner.end_game(GameResult.WIN)
-        looser.end_game(GameResult.LOSE)
+            white = User.get_by_id(match.white_id)
+            black = User.get_by_id(match.black_id)
 
-        # Delete match
-        match.key.delete()
+            if match.is_user_turn(white):
+                looser = white
+                winner = black
+            else:
+                winner = white
+                looser = black
+
+            white.send_message(constants.ERROR_TIMEOUT.format(
+                looser.username,
+                humanfriendly.format_timespan(match.timeout)))
+            black.send_message(constants.ERROR_TIMEOUT.format(
+                looser.username,
+                humanfriendly.format_timespan(match.timeout)))
+
+            winner.end_game(GameResult.WIN)
+            looser.end_game(GameResult.LOSE)
+
+            # Delete match
+            logging.info("Deleting match %s. Last activity: %s", match.key.id(), str(match.last_move_date))
+            match.key.delete()
+
+
+def notify_admins(message):
+
+    admins = list(User.query(User.admin == True).fetch())
+
+    for admin in admins:
+        admin.send_message(message)
